+++ server.js
@@ -63,7 +63,7 @@ const registerRoute = require('./routes/auth/register');
 const loginRoute = require('./routes/auth/login');
 const googleAuth = require('./routes/auth/googleAuth');
 const userOperation = require('./routes/user/user_operations');
-const postsRoute = require('./routes/post');
+const postsRoute = require('./routes/post/postRoute');
 
 /*-----Routes Config End------*/
